Hello World
This is where the program was at the point I decided to upload it to GitHub. It has a few issues with rendering cards and doesn't yet have a proper way of winning/losing or tracking winnings. I plan to fix more of it later and eventually make a decent looking/feeling game out of this basic form.